move FilterSelector class to util/filters, updated tests to cover level index, need to refactor tests a bit more
