+++ src/utils/filter.ts
@@ -1,4 +1,7 @@
+import type { QueryOptions } from '../store/index-level.js';
 import type { EqualFilter, Filter, FilterValue, OneOfFilter, RangeFilter } from '../types/message-types.js';
+
+import { isEmptyObject } from './object.js';
 
 /**
  * A Utility class to help match indexes against filters.
@@ -176,5 +179,154 @@ export class FilterUtility {
       return true;
     };
     return false;
+  }
+}
+
+export class FilterSelector {
+
+  private static hasCursorAndSortProperty(filters: Filter[], queryOptions: QueryOptions):boolean {
+    const { cursor, sortProperty } = queryOptions;
+    return cursor !== undefined &&
+      (sortProperty === 'watermark' || filters.findIndex(filter => Object.keys(filter).map(property => property === sortProperty)) > -1);
+  }
+
+  private static checkRangeFilters(filters: Filter[]): Filter[] {
+    return filters.filter(filter => Object.values(filter).findIndex(filterValue => FilterUtility.isRangeFilter(filterValue)) > -1);
+  }
+
+  private static checkForIdSearches(filters: Filter[]): { searchFilters: Filter[], remainingFilters: Filter[] } {
+    const searchFilters: Filter[] = [];
+    const remainingFilters: Filter[] = [];
+    // next we determine if any of the filters contain a specific identifier such as recordId or permissionsGrantId
+    // if that's the case it's always the only property for the specific filter it's a member of
+    for (const filter of filters) {
+      const { recordId, permissionsGrantId } = filter;
+      // we don't use range filters with these, so either Equality or OneOf filters should be used
+      if (recordId !== undefined && (FilterUtility.isEqualFilter(recordId) || FilterUtility.isOneOfFilter(recordId))) {
+        searchFilters.push({ recordId });
+        continue;
+      }
+
+      if (permissionsGrantId !== undefined && (FilterUtility.isEqualFilter(permissionsGrantId) || FilterUtility.isOneOfFilter(permissionsGrantId))) {
+        searchFilters.push({ permissionsGrantId });
+        continue;
+      }
+
+      remainingFilters.push(filter);
+    }
+
+    return { searchFilters, remainingFilters };
+  }
+
+  private static checkCommonFilters(filters: Filter[]): Filter | undefined {
+    const { schema, contextId, protocol, protocolPath } = this.commonFilters(filters);
+
+    // if we match any of these, we add them to our search filters and return immediately
+    if (contextId !== undefined && FilterUtility.isEqualFilter(contextId)) {
+      // a common contextId exists between all filters
+      // we return this first, as it will likely produce the smallest match set.
+      return { contextId };
+    } else if ( schema !== undefined && FilterUtility.isEqualFilter(schema)) {
+      // a common schema exists between all filters
+      // we return this second, as it will likely produce a sufficiently small match set.
+      return { schema };
+    } else if (protocolPath !== undefined && FilterUtility.isEqualFilter(protocolPath)) {
+      // a common protocol exists between all filters
+      // we return this third, as it will likely produce a sufficiently small match set.
+      return { protocolPath };
+    } else if (protocol !== undefined && FilterUtility.isEqualFilter(protocol)) {
+      // a common protocol exists between all filters
+      // we return this third, as it will likely produce a sufficiently small match set.
+      return { protocol };
+    };
+  }
+
+  /**
+   * Helps select which filter properties are needed to build a filtered query for the LevelDB indexes.
+   *
+   * @param filters the array of filters from an incoming query.
+   * @param queryOptions options associated with the incoming query.
+   * @returns an array of filters to query using. If an empty array is returned, query using the sort property index.
+   */
+  static select(filters: Filter[], queryOptions: QueryOptions): Filter[] {
+
+    // if we have a cursor and this is an EventsQuery (the only query that sorts by watermark), we want to trigger the sortedIndexQuery
+    // we also trigger a sortedIndexQuery if we have a cursor and one of the filters is the same as the sortProperty
+    if (this.hasCursorAndSortProperty(filters, queryOptions)) {
+      return [];
+    }
+
+    // if the number of range filters that exist are equal to the number of filters in the query, we return the range filters.
+    const rangeFilters = this.checkRangeFilters(filters);
+    if (rangeFilters.length === filters.length) {
+      return rangeFilters;
+    }
+
+    const { searchFilters, remainingFilters } = this.checkForIdSearches(filters);
+    // now we determine if the remaining filters array has any common filters.
+    // If there is a match, it's likely best to run a single query against that filter.
+    const commonFilter = this.checkCommonFilters(remainingFilters);
+    if (commonFilter !== undefined) {
+      // the commonFilter was built from the remainingFilters from the checkForIdsSearch function
+      // so we add the returned idFilters array, which could be empty, as well as the remaining common filter.
+      return [ ...searchFilters, commonFilter ];
+    }
+
+    // if we found no common filters, we will attempt to find context, schema, or protocol of each filter
+    const finalFilters: Filter[] = remainingFilters.map(({ contextId, schema, protocol, protocolPath }) => {
+      // if check for single equality filters first in order of most likely to have a smaller set
+      if (contextId !== undefined && FilterUtility.isEqualFilter(contextId)) {
+        return { contextId } as Filter;
+      } else if (schema !== undefined && FilterUtility.isEqualFilter(schema)) {
+        return { schema } as Filter;
+      } else if (protocolPath !== undefined && FilterUtility.isEqualFilter(protocolPath)) {
+        return { protocolPath } as Filter;
+      } else if (protocol !== undefined && FilterUtility.isEqualFilter(protocol)) {
+        return { protocol } as Filter;
+      }
+
+      // check for OneOf filters next
+      if (contextId !== undefined && FilterUtility.isOneOfFilter(contextId)) {
+        return { contextId } as Filter;
+      } else if (schema !== undefined && FilterUtility.isOneOfFilter(schema)) {
+        return { schema } as Filter;
+      } else if (protocolPath !== undefined && FilterUtility.isOneOfFilter(protocolPath)) {
+        return { protocol } as Filter;
+      } else if (protocol !== undefined && FilterUtility.isOneOfFilter(protocol)) {
+        return { protocolPath } as Filter;
+      }
+
+      // we return an empty filter and check for it to
+      return { };
+    });
+
+    // if we have an empty filter, we will query based on the sort property, so we return an empty set of filters.
+    if (finalFilters.findIndex(filter => isEmptyObject(filter)) > -1) {
+      return [];
+    }
+
+    return [ ...finalFilters, ...searchFilters];
+  }
+
+
+  /**
+   * Given an array of filters, it returns a single filter with common property/values amongst all the filters.
+   * If there are no common filters, the filter is empty.
+   */
+  private static commonFilters(filters: Filter[]): Filter {
+    if (filters.length === 0) {
+      return { };
+    }
+    return filters.reduce((prev, current) => {
+      const filterCopy = { ...prev };
+      for (const property in filterCopy) {
+        const filterValue = filterCopy[property];
+        const compareValue = current[property];
+        if (FilterUtility.encodeValue(compareValue) !== FilterUtility.encodeValue(filterValue)) {
+          delete filterCopy[property];
+        }
+      }
+      return filterCopy;
+    });
   }
 }
